requirements.txt
python-dotenv
boto3
pyspark
jupyterlab
pandas
beautifulsoup4
apache-airflow==2.6.2
apache-airflow-providers-google==10.1.0